+++ style_bert_vits2/nlp/english/normalizer.py
@@ -2,6 +2,7 @@ import re
 
 import inflect
 
+
 __INFLECT = inflect.engine()
 __COMMA_NUMBER_PATTERN = re.compile(r"([0-9][0-9\,]+[0-9])")
 __DECIMAL_NUMBER_PATTERN = re.compile(r"([0-9]+\.[0-9]+)")
@@ -119,10 +120,11 @@ def __expand_number(m: re.Match[str]) -> str:
             return __INFLECT.number_to_words(num // 100) + " hundred"  # type: ignore
         else:
             return __INFLECT.number_to_words(
-                num, andword="", zero="oh", group=2  # type: ignore
-            ).replace(
-                ", ", " "
-            )  # type: ignore
+                num,
+                andword="",
+                zero="oh",
+                group=2,  # type: ignore
+            ).replace(", ", " ")  # type: ignore
     else:
         return __INFLECT.number_to_words(num, andword="")  # type: ignore