+++ style_bert_vits2/nlp/japanese/bert_feature.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from collections.abc import Sequence
-from typing import TYPE_CHECKING, Any, Optional, Union
+from typing import TYPE_CHECKING, Any
 
 import numpy as np
 import onnxruntime
@@ -12,6 +12,7 @@ from style_bert_vits2.nlp import bert_models, onnx_bert_models
 from style_bert_vits2.nlp.japanese.g2p import text_to_sep_kata
 from style_bert_vits2.utils import get_onnx_device_options
 
+
 if TYPE_CHECKING:
     import torch
 
@@ -20,7 +21,7 @@ def extract_bert_feature(
     text: str,
     word2ph: list[int],
     device: str,
-    assist_text: Optional[str] = None,
+    assist_text: str | None = None,
     assist_text_weight: float = 0.7,
 ) -> torch.Tensor:
     """
@@ -88,8 +89,8 @@ def extract_bert_feature(
 def extract_bert_feature_onnx(
     text: str,
     word2ph: list[int],
-    onnx_providers: Sequence[Union[str, tuple[str, dict[str, Any]]]],
-    assist_text: Optional[str] = None,
+    onnx_providers: Sequence[str | tuple[str, dict[str, Any]]],
+    assist_text: str | None = None,
     assist_text_weight: float = 0.7,
 ) -> NDArray[Any]:
     """