+++ style_bert_vits2/nlp/japanese/user_dict/__init__.py
@@ -24,6 +24,7 @@ from style_bert_vits2.nlp.japanese.user_dict.part_of_speech_data import (
 )
 from style_bert_vits2.nlp.japanese.user_dict.word_model import UserDictWord, WordTypes
 
+
 # root_dir = engine_root()
 # save_dir = get_save_dir()
 
@@ -45,7 +46,7 @@ compiled_dict_path = (
 
 
 # @mutex_wrapper(mutex_user_dict)
-def _write_to_json(user_dict: Dict[str, UserDictWord], user_dict_path: Path) -> None:
+def _write_to_json(user_dict: dict[str, UserDictWord], user_dict_path: Path) -> None:
     """
     ユーザー辞書ファイルへのユーザー辞書データ書き込み
     Parameters
@@ -166,7 +167,7 @@ def update_dict(
 
 
 # @mutex_wrapper(mutex_user_dict)
-def read_dict(user_dict_path: Path = user_dict_path) -> Dict[str, UserDictWord]:
+def read_dict(user_dict_path: Path = user_dict_path) -> dict[str, UserDictWord]:
     """
     ユーザー辞書の読み出し
     Parameters
@@ -183,7 +184,7 @@ def read_dict(user_dict_path: Path = user_dict_path) -> Dict[str, UserDictWord]:
         return {}
 
     with user_dict_path.open(encoding="utf-8") as f:
-        result: Dict[str, UserDictWord] = {}
+        result: dict[str, UserDictWord] = {}
         for word_uuid, word in json.load(f).items():
             # cost2priorityで変換を行う際にcontext_idが必要となるが、
             # 0.12以前の辞書は、context_idがハードコーディングされていたためにユーザー辞書内に保管されていない
@@ -203,8 +204,8 @@ def _create_word(
     surface: str,
     pronunciation: str,
     accent_type: int,
-    word_type: Optional[WordTypes] = None,
-    priority: Optional[int] = None,
+    word_type: WordTypes | None = None,
+    priority: int | None = None,
 ) -> UserDictWord:
     """
     単語オブジェクトの生成
@@ -256,8 +257,8 @@ def apply_word(
     surface: str,
     pronunciation: str,
     accent_type: int,
-    word_type: Optional[WordTypes] = None,
-    priority: Optional[int] = None,
+    word_type: WordTypes | None = None,
+    priority: int | None = None,
     user_dict_path: Path = user_dict_path,
     compiled_dict_path: Path = compiled_dict_path,
 ) -> str:
@@ -308,8 +309,8 @@ def rewrite_word(
     surface: str,
     pronunciation: str,
     accent_type: int,
-    word_type: Optional[WordTypes] = None,
-    priority: Optional[int] = None,
+    word_type: WordTypes | None = None,
+    priority: int | None = None,
     user_dict_path: Path = user_dict_path,
     compiled_dict_path: Path = compiled_dict_path,
 ) -> None:
@@ -385,7 +386,7 @@ def delete_word(
 
 
 def import_user_dict(
-    dict_data: Dict[str, UserDictWord],
+    dict_data: dict[str, UserDictWord],
     override: bool = False,
     user_dict_path: Path = user_dict_path,
     default_dict_path: Path = default_dict_path,
@@ -449,7 +450,7 @@ def import_user_dict(
     )
 
 
-def _search_cost_candidates(context_id: int) -> List[int]:
+def _search_cost_candidates(context_id: int) -> list[int]:
     for value in part_of_speech_data.values():
         if value.context_id == context_id:
             return value.cost_candidates